+++ CHANGES.md
@@ -5,6 +5,8 @@ This files lists the changes during the lifetime of this project.
 - add period filter to placement page
 - add end date of current placement to colleague list
 - add availability sorting to colleague list
+- (backwards incompatible) change assignment status. new list: LEAD, VACATURE, INGEVULD, AFGEWEZEN
+- changed that assignment phase is computed instead of assigned
 
 ## 2025-09-19
 - update dependencies to latest, including django security release